+++ fabric/src/main/java/top/theillusivec4/veinmining/veinmining/VeinMiningEnchantment.java
@@ -20,6 +20,8 @@ package top.theillusivec4.veinmining.veinmining;
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.enchantment.EnchantmentTarget;
 import net.minecraft.entity.EquipmentSlot;
+import net.minecraft.util.Identifier;
+import net.minecraft.util.registry.Registry;
 import top.theillusivec4.veinmining.config.VeinMiningConfig;
 
 public class VeinMiningEnchantment extends Enchantment {
@@ -62,5 +64,16 @@ public class VeinMiningEnchantment extends Enchantment {
   @Override
   public boolean isAvailableForRandomSelection() {
     return VeinMiningConfig.Enchantment.isRandomlySelectable;
+  }
+
+  @Override
+  protected boolean canAccept(Enchantment other) {
+    Identifier id = Registry.ENCHANTMENT.getId(other);
+
+    if (id != null &&
+        VeinMiningConfig.Enchantment.incompatibleEnchantments.contains(id.toString())) {
+      return false;
+    }
+    return super.canAccept(other);
   }
 }