+++ forge/CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to the format [MCVERSION-MAJORMOD.MAJORAPI.MINOR.PATCH](https://mcforge.readthedocs.io/en/1.16.x/conventions/versioning/).
 
+## [1.16.5-0.8] - 2021.01.31
+### Added
+- Added `incompatibleEnchantments` config option
+
 ## [1.16.5-0.7] - 2021.01.31
 ### Added
 - Added Chinese translation (thanks EnterFor!)